+++ drivers/ethereum/fetch.go
@@ -3,8 +3,19 @@ package ethereum
 import (
 	"context"
 	"github.com/datadaodevs/evm-etl/protos/go/protos/evm/raw"
+	"github.com/datadaodevs/go-service-framework/pool"
 )
 
+// FetchSequence defines the parallelizable steps in the fetch sequence
+func (e *EthereumDriver) FetchSequence(index uint64) map[string]pool.Runner {
+	return map[string]pool.Runner{
+		stageFetchBlock:   e.queueGetBlockByNumber(index),
+		stageFetchReceipt: e.queueGetBlockReceiptsByNumber(index),
+		stageFetchTraces:  e.queueGetBlockTraceByNumber(index),
+	}
+}
+
+// GetChainTipNumber gets the block number of the chaintip
 func (e *EthereumDriver) GetChainTipNumber(ctx context.Context) (uint64, error) {
 	blockNum, err := e.client.EthBlockNumber(ctx)
 	if err != nil {
@@ -15,6 +26,7 @@ func (e *EthereumDriver) GetChainTipNumber(ctx context.Context) (uint64, error) 
 	return blockNum, nil
 }
 
+// getBlockByNumber fetches a full block by number
 func (e *EthereumDriver) getBlockByNumber(ctx context.Context, index uint64) (*raw.Block, error) {
 	block, err := e.client.EthGetBlockByNumber(index)
 	if err != nil {
@@ -25,6 +37,7 @@ func (e *EthereumDriver) getBlockByNumber(ctx context.Context, index uint64) (*r
 	return block, nil
 }
 
+// getBlockTraceByNumber fetches all traces for a given block
 func (e *EthereumDriver) getBlockTraceByNumber(ctx context.Context, index uint64) ([]*raw.CallTrace, error) {
 	traces, err := e.client.DebugTraceBlock(index)
 	if err != nil {
@@ -35,6 +48,7 @@ func (e *EthereumDriver) getBlockTraceByNumber(ctx context.Context, index uint64
 	return traces, nil
 }
 
+// getBlockReceiptsByNumber fetches a set of block receipts for a given block
 func (e *EthereumDriver) getBlockReceiptsByNumber(ctx context.Context, index uint64) ([]*raw.TransactionReceipt, error) {
 	receipts, err := e.client.GetBlockReceipt(index)
 	if err != nil {
@@ -43,4 +57,25 @@ func (e *EthereumDriver) getBlockReceiptsByNumber(ctx context.Context, index uin
 	}
 
 	return receipts, nil
+}
+
+// queueGetBlockTraceByNumber wraps GetBlockTraceByNumber in a queueable Runner func
+func (e *EthereumDriver) queueGetBlockTraceByNumber(index uint64) pool.Runner {
+	return func(ctx context.Context) (interface{}, error) {
+		return e.getBlockTraceByNumber(ctx, index)
+	}
+}
+
+// queueGetBlockByNumber wraps GetBlockByNumber in a queueable Runner func
+func (e *EthereumDriver) queueGetBlockByNumber(index uint64) pool.Runner {
+	return func(ctx context.Context) (interface{}, error) {
+		return e.getBlockByNumber(ctx, index)
+	}
+}
+
+// queueGetBlockReceiptsByNumber wraps GetBlockReceiptsByNumber in a queueable Runner func
+func (e *EthereumDriver) queueGetBlockReceiptsByNumber(index uint64) pool.Runner {
+	return func(ctx context.Context) (interface{}, error) {
+		return e.getBlockReceiptsByNumber(ctx, index)
+	}
 }
